fix(product): magento product previews need to transform stock status
