+++ libs/product/driver/magento/src/models/public_api.ts
@@ -5,7 +5,7 @@ export {
   MagentoBundledProductItem,
   MagentoPriceTypeEnum,
 } from './bundled-product';
-export { MagentoProductPreview } from './product-preview.interface';
+export * from './product-preview.interface';
 export { MagentoSimpleProduct } from './simple-product';
 export * from './configurable-product';
 export { MagentoGetProductResponse } from './get-product-response.interface';
